+++ package/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- added configurability for pan handler
+- added configurability for overlay border radius
+
 ## [0.1.6] - 2024-01-29
 
 - Much smoother transition by transitioning the backdrop image into the tile image.